Consolidate and remove SendComponentWithData
