+++ src/components/fields/AddressField.js
@@ -49,8 +49,9 @@ export default withNavigation(
         this.inputRef.focus()
       );
     }
+
     shouldComponentUpdate(nextProps, nextState) {
-      const isNewAddress = isNewValueForPath(this.props, this.state, 'address');
+      const isNewAddress = isNewValueForPath(nextProps, this.state, 'address');
       const isNewInputValue = isNewValueForPath(
         this.state,
         nextState,
