+++ src/hoc/index.js
@@ -28,6 +28,7 @@ export { default as withRequests } from './withRequests';
 export { default as withRotationForDirection } from './withRotationForDirection';
 export { default as withSafeAreaViewInsetValues } from './withSafeAreaViewInsetValues';
 export { default as withSelectedInput } from './withSelectedInput';
+export { default as withSend } from './withSend';
 export { default as withSendFeedback } from './withSendFeedback';
 export { default as withStatusBarStyle } from './withStatusBarStyle';
 export { default as withTransactionConfirmationScreen } from './withTransactionConfirmationScreen';
