+++ src/redux/gas.js
@@ -177,9 +177,7 @@ const getSelectedGasPrice = (
   };
 };
 
-export const gasClearState = () => () => {
-  clearInterval(getGasPricesInterval);
-};
+export const gasClearState = () => clearInterval(getGasPricesInterval);
 
 // -- Reducer --------------------------------------------------------------- //
 const INITIAL_STATE = {
